Exit on sigint, not client notification
This required not opening lvm and vg handles across the daemon lifetime,
but opening them within each method call. LVM is doing things with signals
that make this a better fit.

Define a context manager to make methods opening lvm and vg as easy as
possible.

Open lvm/vg once on startup, so lvm/vg errors are thrown early.

Signed-off-by: Andy Grover <agrover@redhat.com>