implement vol_copy
Given an existing volume and a new volume name, vol_copy creates the
new volume with the same size as the existing one and copies the contents
to the new volume.

Copy operation does not use shutil, because we are going to be noting
ongoing progress so we need to do the copy loop ourselves.

Signed-off-by: Andy Grover <agrover@redhat.com>