Make name of storageobject pool:name
Storageobject names need to be unique (within type.) Make sure they will
be unique even if we added >1 pools support by prepending the pool name
to the storageobject name, as well as the lv name.

Remove an extra call to BlockStorageObject.

Rename export create/destroy functions to be more in line with API names.

Return vol and pool from export_list, in accordance with API 0.2.

Signed-off-by: Andy Grover <agrover@redhat.com>