Bug fixes for incorrect pool usage.
A previous patch introduced the ability to use thinly
provisioned logical volumes which are used as pools in lvm as
pools in targetd.  The configuration file syntax is vg/thin_lv.
Unfortunately there was a few places where we were incorrectly
using vg/lv_pool for the pool instead of just the vg part.

Signed-off-by: Tony Asleson <tasleson@redhat.com>