Defeature ASYNC and non-thinp copies
We've decided to remove vol_copy support for non-thinp volumes. vol_copy
was the only potentially slow API call, and the reason the API had
ASYNC, which complicated targetd's implementation and usage. We can
now remove ASYNC from the API.

This patch updates API.md to 0.6, which removes ASYNC support. It also
updates targetd and the client example code to no longer support ASYNC.
We still support creating non-thinp volumes, but doing a copy on a
non-thinp volume returns an error.

Signed-off-by: Andy Grover <agrover@redhat.com>