Fix block_pools to handle thinpool LVs
Instead of always reporting the free size of the VG, which will be
wrong if our pool is a thinpool in the VG, if it's a thinpool, try
to calculate how much of it is free by looking at the data_percent
and metadata_percent properties. This isn't quite the format we need
to return, so some munging is involved to come up with an approx. answer.

Signed-off-by: Andy Grover <agrover@redhat.com>