New method: initiator_list(standalone_only=False)
The 'standalone_only' parameter is optional. Default is False which means
all initiators will be included in results.
When standalone_only is True, only return initiator which is not in any
NodeACLGroup.

Return a list of initiator in the format of

    [
        {
            'init_id': str,
            'init_type': str,
        },
    ]

The 'init_id' of result is the iSCSI IQN/NAA/EUI address of initiator.
Example: 'iqn.2000-04.com.example:someone-01'
The 'init_type' is reserved for future use, currently, it is 'iscsi'.

Signed-off-by: Gris Ge <fge@redhat.com>