Check for -1 values when retrieving percentages
targetd was returning -3169126500561983635456 for a thinpool free space
during testing.  This was caused by the api returning -1 when retrieving
the property 'data_percent' & 'metadata_percent'.  This patch checks for
this and returns the full size for the free size if we are unable to get
actual values of consumed percentages.

Signed-off-by: Tony Asleson <tasleson@redhat.com>