nfs.py: Account for new nfs output
It appears that the latest `exportfs -v` has additional output which
caused us to fail during parsing the optionals.  Account for the failures,
but not sure we have all of them.  Need to go back through and check the
code and/or docs for what has changed to ensure we have all cases covered.

Note: We need to improve the error handling.  We incorrectly report that
a method isn't supported when we get a key error.  We got a key error in
the nfs options parsing which was uncaught and thus made it look like the
plugin didn't support the listing of nfs exports, which it certainly does.

Signed-off-by: Tony Asleson <tasleson@redhat.com>