block.py: Check for existing LV name before creating/coping
We were not checking before we created/copied a LV which would cause
the lvm library to return an ambigious error.  Check before doing the
operation until we have a lvm library that has meaningful error codes.
Otherwise we would need to parse the stdout/stderr for error reason which
is less than ideal.

Signed-off-by: Tony Asleson <tasleson@redhat.com>