Try to handle bad config file password
If a user doesn't place a space between the ':' and the
password we simply get a string returned back from the yaml parse.

In this case we will ensure that the returned value is not none and
not simply a string.  Perhaps there is a better way to handle this?

Ref. https://github.com/open-iscsi/targetd/issues/35

Signed-off-by: Tony Asleson <tasleson@redhat.com>