Copy method on ZFS pools
This method works, however there is limitation when using zfs snapshot
and clone: clone is dependant on original dataset and it's snapshot
and ZFS does not offer any way around that (zfs promote allows to
switch parent-child relationship, but doesn't break it).

To make sure user understands implications of using copy on ZFS
it is disabled by default and can be enabled in config file.