main.py: Trap SIGINT signal
If you have targetd running in forground and hit ^C you will exit cleanly.
If targetd is running in background and you send SIGINT, it's ignored.
Not sure why as the documentation would indicate that it should work the
same.

Add signal handler and send SIGINT signal in test.sh script to exit
targetd cleanly.  This way we actually can get the coverage data on
targetd.

Thanks to Daniel Sonck for the suggestion for using handle_request.

Signed-off-by: Tony Asleson <tasleson@redhat.com>